exact-school-backend: adicionando filtros de validação de perfil em todas as rotas e decodificando o token

+++ src/api/controllers/classController.js
@@ -1,11 +1,10 @@
 // C - CONTROLLER
 
-// as operações de manipulação estão dentro do Class (pois ele é um modelo do mongoose que está no outro arquivo)
-// para validar senha bcrypt.compareSync(passwordToCompare, passwordOfDatabase) quando for logar e fazer autenticação
-
 const Class = require('../models/classes')
+const userTypesEnum = require ('../common/enums/userTypes')
 
 const NotFoundError = require('../common/errors/NotFound')
+const UnanthorizedError = require('../common/errors/Unanthorized')
 
 exports.create = async function (req, res, next) {
 
@@ -62,12 +61,18 @@ exports.getById = async function (req, res, next) {
   if (classToGet) res.send(classToGet)
 }
 
-
 exports.update = async function (req, res, next) {
 
   let body = req.body
 
   body.updatedAt = new Date().toISOString()
+
+  if (!(req.authUser.userType === userTypesEnum.ADMIN || req.authUser.id === req.params.id)){
+    const error = new UnanthorizedError()
+    error.httpStatusCode = 401
+    res.status(401).send({ sucess: false, res: 'Sem permissão', status: error.httpStatusCode })
+    return next(error)
+  }
 
   const classToUpdate = await Class.findByIdAndUpdate(req.params.id, { $set: body })
   
@@ -82,6 +87,13 @@ exports.update = async function (req, res, next) {
 }
 
 exports.delete = async function (req, res, next) {
+
+  if (!(req.authUser.userType === userTypesEnum.ADMIN || req.authUser.id === req.params.id)){
+    const error = new UnanthorizedError()
+    error.httpStatusCode = 401
+    res.status(401).send({ sucess: false, res: 'Sem permissão', status: error.httpStatusCode })
+    return next(error)
+  }
 
   const classToDelete = await Class.findByIdAndRemove(req.params.id)
